Move config and weight code into their own files
